Commit of the Masters Degree Software Development Assigment 2 Program
this progam was creaded in Visual basic console using C#. the purpose of it was to answer question tests that were given by the teacher. I Improvised and made a small game out of the questions so that the person can choose which question adn which part they wanted to see.